third_party/WebKit/LayoutTests/platform/win/fast/dom/custom/custom-elements-registry-expected.txt

This is a testharness.js-based test.
PASS window.customElements.define is defined 
PASS window.customElements.define requires two arguments 
Harness: the test ran to completion.
